fix user authentication and move controllers into api folders and restructure files
