+++ src/apis/emails/entities/email.entity.ts
