+++ src/apis/users/dtos/login-user.dto.ts
@@ -1,0 +1,10 @@
+import { IsEmail, IsString } from 'class-validator';
+
+export class LoginUserDto {
+
+  @IsEmail()
+  email: string;
+
+  @IsString()
+  password: string;
+}
