+++ src/apis/users/users.module.ts
@@ -6,9 +6,9 @@ import { AuthService } from './auth.service';
 import { User } from './user.entity';
 import { APP_INTERCEPTOR } from '@nestjs/core';
 import { CurrentUserInterceptor } from './interceptors/current-user.interceptors';
-import { WalletsService } from 'src/wallets/wallets.service';
-import { Wallet } from 'src/wallets/wallet.entity';
-import { EmailsService } from 'src/emails/emails.service';
+import { WalletsService } from 'src/apis/wallets/wallets.service';
+import { Wallet } from 'src/apis/wallets/wallet.entity';
+import { EmailsService } from 'src/apis/emails/emails.service';
 import { JwtService } from '@nestjs/jwt';
 
 @Module({
