+++ src/apis/wallets/wallets.module.ts
