+++ src/cron-jobs/cron-jobs.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { CronJobsService } from './cron-jobs.service';
 import { CronJobsController } from './cron-jobs.controller';
-import { EmailsService } from 'src/emails/emails.service';
+import { EmailsService } from 'src/apis/emails/emails.service';
 import { JwtService } from '@nestjs/jwt';
-import { UsersService } from 'src/users/users.service';
+import { UsersService } from 'src/apis/users/users.service';
 import { User } from 'src/entities';
 import { TypeOrmModule } from '@nestjs/typeorm';
