--- src/cron-jobs/entities/cron-job.entity.ts
@@ -1,1 +1,0 @@
-export class CronJob {}
